Seperate server and app for testing
