+++ server/models/reservations.js
@@ -4,11 +4,11 @@ const models = require('../../database/dbModels');
 mongoose.connect('mongodb://localhost/restaurants');
 
 //searches a DB Model for particular id, and invokes callback on it
-const findInDb = function findInDb(model, id, callback){
+const findInDb = function findInDb(model, id, res, callback){
 
   model.find({id:id}, (err, result)=>{
-    if (err){
-      console.log('Error retrieving data', err);
+    if (err || result[0] === undefined){
+      console.log('Error retrieving data', err || 'no results from DB');
       res.status(500).send(err);
       return;
     }
@@ -17,13 +17,13 @@ const findInDb = function findInDb(model, id, callback){
 }
 
 //Query db for specific restaurant availability times
-const availability = function availability(id, callback){
-  findInDb(models.availability, id, callback);
+const availability = function availability(id, res, callback){
+  findInDb(models.availability, id, res, callback);
 }
 
 //Query db for specific restaurant booking times
-const bookingsCount = function bookingsCount(id, callback){
-  findInDb(models.bookings, id, callback);
+const bookingsCount = function bookingsCount(id, res, callback){
+  findInDb(models.bookings, id, res, callback);
 }
 
 module.exports.availability = availability;
