+++ server/restaurantRouter.js
@@ -6,7 +6,7 @@ const reservations = require('./models/reservations');
 //Route requests for specific restaurant availability
 router.get('/:id/reservations', (req,res)=>{
 
-  reservations.availability(req.params.id, (data)=>{
+  reservations.availability(req.params.id, res, (data)=>{
     //send back table availability times
     res.send(data.availability); 
   });
@@ -14,7 +14,7 @@ router.get('/:id/reservations', (req,res)=>{
 
 //Route requests for specific restaurant bookings count
 router.get('/:id/bookings', (req,res)=>{
-  reservations.bookings(req.params.id, (data)=>{
+  reservations.bookings(req.params.id, res, (data)=>{
     //send back booking times
     res.send(data.bookings);
   });
